refact: carousel 분리
